requirements.txt
bitarray==2.7.3
cffi==1.15.1
cryptography==40.0.2
docx2python==2.6.3
docx2txt==0.8
et-xmlfile==1.1.0
jdcal==1.4.1
lxml==4.9.2
msoffcrypto-tool==5.0.1
numpy==1.19.3
olefile==0.46
openpyxl==2.6.2
pdfreader==0.1.12
Pillow==9.5.0
pkg_resources==0.0.0
pycparser==2.21
pycryptodome==3.17
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2023.3
rarfile==4.0
six==1.16.0
typing_extensions==4.5.0
tzdata==2023.3
XlsxWriter==3.1.0
zipfile36==0.1.3
